requirements.txt
flask==3.0.2
pandas==2.2.1
python-dotenv==1.0.1
matplotlib==3.8.3
numpy==1.26.4
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.5 